# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "X-Realty Estate API",
        "description": "API endpoints for real estate management",
        "contact": {
            "name": "X-Realty Estate API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/google-maps/geocode": {
            "post": {
                "tags": [
                    "Google Maps"
                ],
                "summary": "Geocode an address using Google Maps API",
                "operationId": "efbd679940737ff848185b14481d4c62",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "address"
                                ],
                                "properties": {
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St, Sydney NSW"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Geocode result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "lat": {
                                            "type": "number"
                                        },
                                        "lng": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unable to geocode address",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/google-maps/autocomplete": {
            "get": {
                "tags": [
                    "Google Maps"
                ],
                "summary": "Get Google Places Autocomplete suggestions for an address input",
                "operationId": "007e2869e7201fbdb278749bbd032c39",
                "parameters": [
                    {
                        "name": "input",
                        "in": "query",
                        "description": "Partial address or place input",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "123 Main St, Syd"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Lat,lng bias (optional)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "-33.8688,151.2093"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Radius in meters (optional)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 50000
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Autocomplete results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "predictions": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input or API error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/payments/subscribe": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Subscribe the authenticated user to a plan",
                "operationId": "12c73d87e5667dcb72ced3c029f587c8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plan"
                                ],
                                "properties": {
                                    "plan": {
                                        "type": "string",
                                        "example": "premium"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "pm_card_visa"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscription created"
                    },
                    "400": {
                        "description": "Error"
                    }
                }
            }
        },
        "/payments/one-time": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Create a one-time payment for a listing upgrade",
                "operationId": "97fdc16fc27c77defae0e63a2c8a73a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "example": 99
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PaymentIntent client secret returned"
                    },
                    "400": {
                        "description": "Error"
                    }
                }
            }
        },
...
    },
...
}